Add files via upload
This repository contains the code used to estimate the Bayesian Model Averaging (BMA) models presented in Chapter 5 of my master's thesis. In addition to the main scripts, it includes supplementary routines designed to explore different outputs provided by the BMA package in Stata, some of which were not explicitly reported in the final version of the thesis.

This repository may be useful as a reference for reproducing the main results from Chapter 5, exploring the implementation of the bmaregress command and its related functions (bmagraph, bmastats, among others), and reviewing examples of how to process BMA outputs and perform complementary analyses not included in the thesis manuscript.